+++ Keas.Mvc/ClientApp/components/People/PeopleContainer.tsx
@@ -36,7 +36,7 @@ export default class PeopleContainer extends React.Component<{}, IState> {
     }
 
     public async componentDidMount() {
-        const people = await this.context.fetch(`/api/${this.context.team.slug}/search/peopleList/`);
+        const people = await this.context.fetch(`/api/${this.context.team.slug}/people/listPeople/`);
         const tags = await this.context.fetch(`/api/${this.context.team.slug}/tags/listTags`);
 
         this.setState({ loading: false, people, tags });
@@ -170,7 +170,7 @@ export default class PeopleContainer extends React.Component<{}, IState> {
         }
         person.teamId = this.context.team.id;
         // any errors here are caught in CreatePerson
-        person = await this.context.fetch(`/api/${this.context.team.slug}/search/createPerson`, {
+        person = await this.context.fetch(`/api/${this.context.team.slug}/people/createPerson`, {
             body: JSON.stringify(person),
             method: "POST"
         });
@@ -200,7 +200,7 @@ export default class PeopleContainer extends React.Component<{}, IState> {
         }
 
         const updated: IPerson = await this.context.fetch(
-            `/api/${this.context.team.slug}/people/update`,
+            `/api/${this.context.team.slug}/peopleAdmin/update`,
             {
                 body: JSON.stringify(person),
                 method: "POST"
@@ -230,7 +230,7 @@ export default class PeopleContainer extends React.Component<{}, IState> {
         }
 
         const deleted: IPerson = await this.context.fetch(
-            `/api/${this.context.team.slug}/people/delete`,
+            `/api/${this.context.team.slug}/peopleAdmin/delete`,
             {
                 body: JSON.stringify(person),
                 method: "POST"
